Add FactoryMethods

+++ Hierarchy/Classes/AgriCultural.cs
@@ -9,7 +9,7 @@ namespace Hierarchy.Classes
     public class AgriCultural: CommercialAviation, IAtomization
     {
         public Atomization Atomization { get; private set; }
-        public AgriCultural(string name, int carrying, int fuelcapacity, int range, int crew,
+        public AgriCultural(string name, int carrying, double fuelcapacity, double range, int crew,
             string regnumber, Atomization atomization)
             : base(name, carrying, fuelcapacity, range, crew, regnumber)
         {
