+++ Hierarchy/Classes/Aircraft.cs
@@ -10,10 +10,10 @@ namespace Hierarchy.Classes
     {
         public string Name { get; private set; }
         public int Carrying { get; private set; }
-        public int FuelCapacity { get; private set; }
-        public int Range { get; private set; }
+        public double FuelCapacity { get; private set; }
+        public double Range { get; private set; }
         public int Crew { get; private set; }
-        public Aircraft(string name, int carrying, int fuelcapacity, int range, int crew)
+        public Aircraft(string name, int carrying, double fuelcapacity, double range, int crew)
         {
             Name = name;
             Carrying = carrying;
