+++ Hierarchy/Classes/MilitaryAviation.cs
@@ -10,7 +10,7 @@ namespace Hierarchy.Classes
     {
         public int BoardNumber { get; private set; }
         public WeaponType WeaponType { get; private set; }
-        public MilitaryAviation(string name, int carrying, int fuelcapacity, int range, int crew,
+        public MilitaryAviation(string name, int carrying, double fuelcapacity, double range, int crew,
             int boardnumber, WeaponType weapontype) : base(name, carrying, fuelcapacity, range, crew)
         {
             BoardNumber = boardnumber;
